Create MediaLoader singleton
